+++ src/OddJob/NullLoggerFactory.cs
@@ -6,6 +6,8 @@ namespace OddJob
 #pragma warning disable SA1600 // Elements should be documented
     internal sealed class NullLoggerFactory : ILoggerFactory
     {
+        public static readonly NullLoggerFactory Instance = new NullLoggerFactory();
+
         public void Dispose()
         {
         }