+++ src/OddJob/Schedules/IEvery.cs
@@ -1,0 +1,9 @@
+namespace OddJob.Schedules
+{
+    /// <summary>
+    /// Defines an interface for building schedules that run on a fix period.
+    /// </summary>
+    public interface IEvery
+    {
+    }
+}